app/routes/estimate_age.py
from fastapi import APIRouter, HTTPException
from deepface import DeepFace
from pydantic import BaseModel
from app.models.schemas import AgeEstimationResponse
import shutil
import tempfile
import os
import cv2
import numpy as np

router = APIRouter()

class AgeRequest(BaseModel):
    image: str # Base64 or dummy for testing
    declared_age: int

def analyze_face(image_data: str):
    """Function exposed for mocking in tests."""
    # In a real scenario, this would decode base64. 
    # For now, we'll return a placeholder that gets patched in tests.
    return {"estimated_age": 30, "face_confidence": 0.9}

@router.post("/estimate-age", response_model=AgeEstimationResponse)
async def estimate_age(request: AgeRequest):
    try:
        # For actual DeepFace call, we'd need a real file.
        # But we use analyze_face to allow patching.
        res = analyze_face(request.image)
        estimated_age = res['estimated_age']
        
        match = abs(estimated_age - request.declared_age) <= 5
        confidence = res.get('face_confidence', 0.9) 

        return AgeEstimationResponse(
            estimated_age=int(estimated_age),
            declared_age=request.declared_age,
            match=match,
            confidence=confidence
        )
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))
