requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
faster-whisper
groq
google-generativeai
deepface
fer
mediapipe
opencv-python
tf-keras
pytest
httpx
pydantic
pydantic-settings
numpy
av
accelerate
bitsandbytes
transformers>=4.42.0
torch
torchvision
